Update transport tests to 6.1.2
