+++ src/NServiceBus.SqlServer.TransportTests/App_Packages/NSB.TransportTests.6.1.2/When_on_error_throws.cs
