+++ src/NServiceBus.SqlServer.TransportTests/App_Packages/NSB.TransportTests.6.1.2/When_sending_from_on_error.cs
