+++ src/NServiceBus.SqlServer.TransportTests/NServiceBus.SqlServer.TransportTests.csproj
@@ -32,7 +32,7 @@
   </PropertyGroup>
   <ItemGroup>
     <Reference Include="NServiceBus.Core, Version=6.0.0.0, Culture=neutral, PublicKeyToken=9fc386479f8a226c, processorArchitecture=MSIL">
-      <HintPath>..\packages\NServiceBus.6.1.1\lib\net452\NServiceBus.Core.dll</HintPath>
+      <HintPath>..\packages\NServiceBus.6.1.2\lib\net452\NServiceBus.Core.dll</HintPath>
       <Private>True</Private>
     </Reference>
     <Reference Include="nunit.framework, Version=3.4.1.0, Culture=neutral, PublicKeyToken=2638cd05610744eb, processorArchitecture=MSIL">
@@ -60,22 +60,23 @@
     <None Include="packages.config" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="App_Packages\NSB.TransportTests.6.1.1\IConfigureTransportInfrastructure.cs" />
-    <Compile Include="App_Packages\NSB.TransportTests.6.1.1\NServiceBusTransportTest.cs" />
-    <Compile Include="App_Packages\NSB.TransportTests.6.1.1\TransportConfigurationResult.cs" />
-    <Compile Include="App_Packages\NSB.TransportTests.6.1.1\TypeScanner.cs" />
-    <Compile Include="App_Packages\NSB.TransportTests.6.1.1\When_failure_happens_after_send.cs" />
-    <Compile Include="App_Packages\NSB.TransportTests.6.1.1\When_message_is_available.cs" />
-    <Compile Include="App_Packages\NSB.TransportTests.6.1.1\When_on_error_throws.cs" />
-    <Compile Include="App_Packages\NSB.TransportTests.6.1.1\When_on_message_throws.cs" />
-    <Compile Include="App_Packages\NSB.TransportTests.6.1.1\When_on_message_throws_after_delayed_retry.cs" />
-    <Compile Include="App_Packages\NSB.TransportTests.6.1.1\When_on_message_throws_after_immediate_retry.cs" />
-    <Compile Include="App_Packages\NSB.TransportTests.6.1.1\When_requesting_immediate_retry.cs" />
-    <Compile Include="App_Packages\NSB.TransportTests.6.1.1\When_scope_dispose_throws.cs" />
-    <Compile Include="App_Packages\NSB.TransportTests.6.1.1\When_sending_from_on_error.cs" />
-    <Compile Include="App_Packages\NSB.TransportTests.6.1.1\When_transaction_level_TransactionScope.cs" />
-    <Compile Include="App_Packages\NSB.TransportTests.6.1.1\When_user_aborts_processing.cs" />
-    <Compile Include="App_Packages\NSB.TransportTests.6.1.1\When_using_non_durable_delivery.cs" />
+    <Compile Include="App_Packages\NSB.TransportTests.6.1.2\IConfigureTransportInfrastructure.cs" />
+    <Compile Include="App_Packages\NSB.TransportTests.6.1.2\NServiceBusTransportTest.cs" />
+    <Compile Include="App_Packages\NSB.TransportTests.6.1.2\TransportConfigurationResult.cs" />
+    <Compile Include="App_Packages\NSB.TransportTests.6.1.2\TypeScanner.cs" />
+    <Compile Include="App_Packages\NSB.TransportTests.6.1.2\When_failure_happens_after_send.cs" />
+    <Compile Include="App_Packages\NSB.TransportTests.6.1.2\When_message_is_available.cs" />
+    <Compile Include="App_Packages\NSB.TransportTests.6.1.2\When_on_error_throws.cs" />
+    <Compile Include="App_Packages\NSB.TransportTests.6.1.2\When_on_message_throws.cs" />
+    <Compile Include="App_Packages\NSB.TransportTests.6.1.2\When_on_message_throws_after_delayed_retry.cs" />
+    <Compile Include="App_Packages\NSB.TransportTests.6.1.2\When_on_message_throws_after_immediate_retry.cs" />
+    <Compile Include="App_Packages\NSB.TransportTests.6.1.2\When_requesting_immediate_retry.cs" />
+    <Compile Include="App_Packages\NSB.TransportTests.6.1.2\When_scope_complete_throws.cs" />
+    <Compile Include="App_Packages\NSB.TransportTests.6.1.2\When_scope_throws_after_successful_message_processing.cs" />
+    <Compile Include="App_Packages\NSB.TransportTests.6.1.2\When_sending_from_on_error.cs" />
+    <Compile Include="App_Packages\NSB.TransportTests.6.1.2\When_transaction_level_TransactionScope.cs" />
+    <Compile Include="App_Packages\NSB.TransportTests.6.1.2\When_user_aborts_processing.cs" />
+    <Compile Include="App_Packages\NSB.TransportTests.6.1.2\When_using_non_durable_delivery.cs" />
     <Compile Include="ConfigureSqlServerTransportInfrastructure.cs" />
   </ItemGroup>
   <ItemGroup />
