+++ src/NServiceBus.SqlServer.TransportTests/packages.config
@@ -1,6 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="NServiceBus" version="6.1.1" targetFramework="net452" />
-  <package id="NServiceBus.TransportTests.Sources" version="6.1.1" targetFramework="net452" />
+  <package id="NServiceBus" version="6.1.2" targetFramework="net452" />
+  <package id="NServiceBus.TransportTests.Sources" version="6.1.2" targetFramework="net452" />
   <package id="NUnit" version="3.4.1" targetFramework="net452" />
 </packages>
